refactor: 변경된 이름에 따른 경로 수정
