+++ frontend/src/pages/EnterPhoneNumber/index.tsx
@@ -3,7 +3,7 @@ import { Container, IconWrapper, PageContainer, PrivacyBox, Title } from './styl
 import { IoIosArrowBack } from 'react-icons/io';
 import { useNavigate } from 'react-router-dom';
 import { ROUTER_PATH } from '../../constants';
-import { useRedirectRegisterPage } from '../../hooks/useCafeId';
+import { useRedirectRegisterPage } from '../../hooks/useRedirectRegisterPage';
 
 const EnterPhoneNumber = () => {
   useRedirectRegisterPage();
